Fixed buttons on Sign forms
The buttons are now visable.  I edited the links shared erb file to add style to the buttons. I then had to edit the index, sign in, and sign up page for the buttons to show up. I noticed I need the forget passord to grab the template to work. Will contiune to make the forms stylized.